chore: resolve test chain mocking
