+++ src/indexers/protocol-metric-indexer.ts
@@ -1,12 +1,12 @@
 import { getDataMapper } from '../aws/dynamodb.utils';
 import { ProtocolMetricSnapshot } from '../aws/models/protocol-metric-snapshot.model';
-import { SUPPORTED_CHAINS } from '../chains/chain';
+import { getSupportedChains } from '../chains/chains.utils';
 import { MetricType } from '../metrics/enums/metric-type';
 import { getChainMetrics } from '../metrics/metrics.utils';
 
 export const indexProtocolMetrics = async () => {
   const mapper = getDataMapper();
-  const metric = await getChainMetrics(SUPPORTED_CHAINS);
+  const metric = await getChainMetrics(getSupportedChains());
   const metricSnapshot = Object.assign(new ProtocolMetricSnapshot(), { ...metric, type: MetricType.protocol });
 
   try {
