+++ src/indexers/vault-definition-indexer.spec.ts
@@ -4,10 +4,10 @@ import * as gqlGenT from '@badger-dao/sdk/lib/graphql/generated/badger';
 import graphVaults from '@badger-dao/sdk-mocks/generated/ethereum/graph/loadSetts.json';
 import registryVaults from '@badger-dao/sdk-mocks/generated/ethereum/vaults/loadVaults.json';
 
-import { SUPPORTED_CHAINS } from '../chains/chain';
+import { getSupportedChains } from '../chains/chains.utils';
 import { TEST_CURRENT_TIMESTAMP } from '../test/constants';
-import { mockQuery } from '../test/mocks.utils';
-import { mockSupportedChains, setupVaultsCoumpoundDDB } from '../test/tests.utils';
+import { mockQuery, setupMockChain } from '../test/mocks.utils';
+import { setupVaultsCoumpoundDDB } from '../test/tests.utils';
 import { captureVaultData } from './vault-definition-indexer';
 
 describe('vault-definition-indexer', () => {
@@ -35,7 +35,7 @@ describe('vault-definition-indexer', () => {
         return { sett: <gqlGenT.SettQuery['sett']>graphVault, __typename: 'Query' };
       });
 
-      await mockSupportedChains();
+      await setupMockChain();
 
       put = jest.spyOn(DataMapper.prototype, 'put').mockImplementation();
     });
@@ -54,7 +54,7 @@ describe('vault-definition-indexer', () => {
 
       await captureVaultData();
 
-      expect(consoleWarnMock.mock.calls.length).toBe(SUPPORTED_CHAINS.length);
+      expect(consoleWarnMock.mock.calls.length).toBe(getSupportedChains().length);
       expect(put.mock.calls.length).toBe(0);
     });
