+++ src/vaults/vaults.utils.spec.ts
@@ -1,34 +1,21 @@
-import {
-  BadgerGraph,
-  gqlGenT,
-  Protocol,
-  VaultBehavior,
-  VaultDTO,
-  VaultsService,
-  VaultType,
-  VaultVersion,
-} from '@badger-dao/sdk';
+import { Protocol, VaultBehavior, VaultDTO, VaultType, VaultVersion } from '@badger-dao/sdk';
 import { BadRequest } from '@tsed/exceptions';
 import { ethers } from 'ethers';
 
 import { Chain } from '../chains/config/chain.config';
-import { ChainVaults } from '../chains/vaults/chain.vaults';
 import { TOKENS } from '../config/tokens.config';
 import { BouncerType } from '../rewards/enums/bouncer-type.enum';
 import { SourceType } from '../rewards/enums/source-type.enum';
 import * as rewardsUtils from '../rewards/rewards.utils';
-import { MOCK_VAULT, MOCK_VAULT_DEFINITION, TEST_ADDR } from '../test/constants';
+import { MOCK_VAULT, MOCK_VAULT_DEFINITION } from '../test/constants';
 import { mockQuery, setupMockChain } from '../test/mocks.utils';
-// import { randomSnapshot } from "../test/tests.utils";
+import { randomSnapshot } from '../test/tests.utils';
 import { TokenNotFound } from '../tokens/errors/token.error';
 import { fullTokenMockMap } from '../tokens/mocks/full-token.mock';
-import { vaultsGraphSdkMapMock } from './mocks/vaults-graph-sdk-map.mock';
-import { vaultsHarvestsSdkMock } from './mocks/vaults-harvests-sdk.mock';
 import {
   defaultVault,
   estimateDerivativeEmission,
   getCachedVault,
-  getVaultHarvestsOnChain,
   getVaultPerformance,
   getVaultTokenPrice,
 } from './vaults.utils';
@@ -116,23 +103,23 @@ describe('vaults.utils', () => {
       });
     });
 
-    // describe("a cached vault exists", () => {
-    //   it("returns the vault", async () => {
-    //     const snapshot = randomSnapshot(MOCK_VAULT_DEFINITION);
-    //     mockQuery([snapshot]);
-    //     const cached = await getCachedVault(chain, MOCK_VAULT_DEFINITION);
-    //     const expected = await defaultVault(chain, MOCK_VAULT_DEFINITION);
-    //     expected.available = snapshot.available;
-    //     expected.pricePerFullShare = snapshot.balance / snapshot.totalSupply;
-    //     expected.balance = snapshot.balance;
-    //     expected.value = snapshot.value;
-    //     expected.boost = {
-    //       enabled: snapshot.boostWeight > 0,
-    //       weight: snapshot.boostWeight
-    //     };
-    //     expect(cached).toMatchObject(expected);
-    //   });
-    // });
+    describe('a cached vault exists', () => {
+      it('returns the vault', async () => {
+        const snapshot = randomSnapshot(MOCK_VAULT_DEFINITION);
+        mockQuery([snapshot]);
+        const cached = await getCachedVault(chain, MOCK_VAULT_DEFINITION);
+        const expected = await defaultVault(chain, MOCK_VAULT_DEFINITION);
+        expected.available = snapshot.available;
+        expected.pricePerFullShare = snapshot.balance / snapshot.totalSupply;
+        expected.balance = snapshot.balance;
+        expected.value = snapshot.value;
+        expected.boost = {
+          enabled: snapshot.boostWeight > 0,
+          weight: snapshot.boostWeight,
+        };
+        expect(cached).toMatchObject(expected);
+      });
+    });
   });
 
   describe('getVaultTokenPrice', () => {
@@ -228,35 +215,5 @@ describe('vaults.utils', () => {
         expect(estimateDerivativeEmission(compound, emission, compoundEmission)).toEqual(expected);
       },
     );
-  });
-
-  describe('getVaultHarvestsOnChain', () => {
-    function setupOnChainHarvests() {
-      // eslint-disable-next-line
-      jest.spyOn(VaultsService.prototype, 'listHarvests').mockImplementation(async ({ address }): Promise<any> => {
-        return vaultsHarvestsSdkMock[address];
-      });
-      /* eslint-disable @typescript-eslint/ban-ts-comment */
-      jest
-        .spyOn(BadgerGraph.prototype, 'loadSett')
-        .mockImplementation(async ({ id, block }): Promise<gqlGenT.SettQuery> => {
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          return vaultsGraphSdkMapMock[`${id.toLowerCase()}-${(block || {}).number || 0}`];
-        });
-    }
-
-    it('returns vaults harvests with apr', async () => {
-      setupOnChainHarvests();
-      expect(await getVaultHarvestsOnChain(chain, TEST_ADDR)).toMatchSnapshot();
-    });
-
-    it('returns empty harvests for unknown vault', async () => {
-      setupOnChainHarvests();
-      jest.spyOn(ChainVaults.prototype, 'getVault').mockImplementation(async (_) => {
-        throw new Error('Missing Vault');
-      });
-      await expect(getVaultHarvestsOnChain(chain, '0x000000000000')).rejects.toThrow(Error);
-    });
   });
 });
